+++ virtualizarr/tests/test_readers/conftest.py
@@ -17,6 +17,15 @@ except ModuleNotFoundError:
 
 
 @pytest.fixture
+def geotiff_file(tmp_path: Path) -> str:
+    """Create a NetCDF4 file with air temperature data."""
+    filepath = tmp_path / "air.tif"
+    with xr.tutorial.open_dataset("air_temperature") as ds:
+        ds.isel(time=0).rio.to_raster(filepath)
+    return str(filepath)
+
+
+@pytest.fixture
 def empty_chunks_hdf5_file(tmpdir):
     ds = xr.Dataset({"data": []})
     filepath = f"{tmpdir}/empty_chunks.nc"
